pyproject.toml
[tool.poetry]
name = "semantic-search-transformers"
version = "0.1.0"
description = "Semantic Search and Information Retrieval with Transformers - Working Examples"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.36.0"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
torch = "^2.1.0"
rank-bm25 = "^0.2.2"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
optimum = "^1.16.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.2"
scikit-learn = "^1.3.0"
datasets = "^2.15.0"
chromadb = {version = "^0.4.0", optional = true}
openai = "^1.0.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.6"

[tool.poetry.extras]
chromadb = ["chromadb"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
